+++ tests/test_grid.py
@@ -3,7 +3,7 @@
 import numpy as np
 
 import tidy3d as td
-from tidy3d.components.grid import Coords, FieldGrid, YeeGrid, Grid
+from tidy3d.components.grid.grid import Coords, FieldGrid, YeeGrid, Grid
 from tidy3d.components.types import TYPE_TAG_STR
