--- tidy3d/components/__init__.py
@@ -1,54 +1,0 @@
-""" Imports all tidy3d """
-
-# grid
-from .grid import Grid, Coords, GridSpec, UniformGrid, CustomGrid, AutoGrid
-
-# geometry
-from .geometry import Box, Sphere, Cylinder, PolySlab
-from .geometry import Geometry, GeometryGroup
-
-# medium
-from .medium import Medium, PoleResidue, Sellmeier, Debye, Drude, Lorentz, AnisotropicMedium, PEC
-from .medium import AbstractMedium, DispersiveMedium, PECMedium
-
-# structure
-from .structure import Structure
-
-# mode
-from .mode import ModeSpec
-
-# source
-from .source import GaussianPulse, ContinuousWave
-from .source import UniformCurrentSource, PlaneWave, ModeSource, PointDipole
-from .source import GaussianBeam, AstigmaticGaussianBeam
-
-# monitor
-from .monitor import FreqMonitor, TimeMonitor, FieldMonitor, FieldTimeMonitor
-from .monitor import Monitor, FluxMonitor, FluxTimeMonitor, ModeMonitor
-from .monitor import ModeSolverMonitor, PermittivityMonitor
-from .monitor import Near2FarAngleMonitor, Near2FarCartesianMonitor, Near2FarKSpaceMonitor
-
-# simulation
-from .simulation import Simulation
-
-# data
-from .data import ScalarFieldDataArray, ScalarModeFieldDataArray, ScalarFieldTimeDataArray
-from .data import ModeAmpsDataArray, ModeIndexDataArray
-from .data import FluxDataArray, FluxTimeDataArray
-from .data import AbstractNear2FarData
-from .data import Near2FarAngleDataArray, Near2FarCartesianDataArray, Near2FarKSpaceDataArray
-
-from .data import FieldData, FieldTimeData, PermittivityData
-from .data import FluxData, FluxTimeData
-from .data import ModeData, ModeSolverData
-from .data import Near2FarAngleData, Near2FarCartesianData, Near2FarKSpaceData
-from .data import Near2FarSurface, RadiationVectors
-
-from .data import SimulationData
-from .data import DATA_TYPE_MAP
-
-# boundary
-from .boundary import BoundarySpec, Boundary, BoundaryEdge, BoundaryEdgeType
-from .boundary import BlochBoundary, Symmetry, Periodic, PECBoundary, PMCBoundary
-from .boundary import PML, StablePML, Absorber, PMLParams, AbsorberParams, PMLTypes
-from .boundary import DefaultPMLParameters, DefaultStablePMLParameters, DefaultAbsorberParameters
