--- tidy3d/components/data/__init__.py
@@ -1,15 +1,0 @@
-"""Imports all user-facing data objects."""
-from .data_array import ScalarFieldDataArray, ScalarModeFieldDataArray, ScalarFieldTimeDataArray
-from .data_array import ModeAmpsDataArray, ModeIndexDataArray
-from .data_array import FluxDataArray, FluxTimeDataArray
-from .data_array import Near2FarAngleDataArray, Near2FarCartesianDataArray, Near2FarKSpaceDataArray
-
-from .monitor_data import FieldData, FieldTimeData, PermittivityData
-from .monitor_data import FluxData, FluxTimeData
-from .monitor_data import ModeData, ModeSolverData
-
-from .monitor_data_n2f import AbstractNear2FarData
-from .monitor_data_n2f import Near2FarAngleData, Near2FarCartesianData, Near2FarKSpaceData
-from .near2far import Near2FarSurface, RadiationVectors
-
-from .sim_data import SimulationData, DATA_TYPE_MAP
