+++ tidy3d/components/data/monitor_data.py
@@ -10,7 +10,7 @@ import pydantic as pd
 from ..base import TYPE_TAG_STR, Tidy3dBaseModel
 from ..types import Axis, Coordinate
 from ..boundary import Symmetry
-from ..grid import Grid
+from ..grid.grid import Grid
 from ..validators import enforce_monitor_fields_present
 from ..monitor import MonitorType, FieldMonitor, FieldTimeMonitor, ModeSolverMonitor
 from ..monitor import ModeMonitor, FluxMonitor, FluxTimeMonitor, PermittivityMonitor
