--- tidy3d/components/grid/__init__.py
@@ -1,5 +1,0 @@
-""" Imports all tidy3d """
-
-# grid
-from .grid import Grid, Coords, Coords1D, YeeGrid, FieldGrid
-from .grid_spec import GridSpec, UniformGrid, CustomGrid, AutoGrid
