+++ tidy3d/components/simulation.py
@@ -15,7 +15,8 @@ from .validators import assert_unique_names, assert_objects_in_sim_bounds
 from .validators import validate_mode_objects_symmetry
 from .geometry import Box
 from .types import Ax, Shapely, FreqBound, Axis, annotate_type
-from .grid import Coords1D, Grid, Coords, GridSpec, UniformGrid
+from .grid.grid import Coords1D, Grid, Coords
+from .grid.grid_spec import GridSpec, UniformGrid
 from .medium import Medium, MediumType, AbstractMedium, PECMedium
 from .boundary import BoundarySpec, Symmetry, BlochBoundary, PECBoundary, PMCBoundary
 from .boundary import PML, StablePML, Absorber
