+++ tidy3d/plugins/dispersion/fit.py
@@ -12,7 +12,7 @@ from rich.progress import Progress
 from pydantic import Field, validator
 
 from ...components.base import Tidy3dBaseModel
-from ...components import PoleResidue, AbstractMedium
+from ...components.medium import PoleResidue, AbstractMedium
 from ...constants import C_0, HBAR, MICROMETER
 from ...components.viz import add_ax_if_none
 from ...components.types import Ax, ArrayLike
